PHP: review #76
I am not sure if the discovery rules makes sense. I don't know
where the `false` comes from. Wouldn't it be smarter to check for sth. like
`.argument.code("CV")`?

But because of https://github.com/testable-eu/sast-tp-framework/issues/10 (#10) I can't test this with the
CPG used in the framework.